wip: split gui app surfaces
